Separate out FstSegmenter from ThaiSegmenter
